+++ product-service/src/helpers/buildResponse.ts
