+++ product-service/src/helpers/isProductValid.ts
